remove unused createJwtRquest method
